Add support for partials in FunctionBuilder
