+++ tests/test_funcutils_fb_py3.py
@@ -1,10 +1,15 @@
 
 import inspect
+import functools
 from collections import defaultdict
 
 import pytest
 
-from boltons.funcutils import wraps, FunctionBuilder
+from boltons.funcutils import wraps, FunctionBuilder, update_wrapper
+
+
+def wrappable_varkw_func(a, b, **kw):
+    return a, b
 
 
 def pita_wrap(flag=False):
@@ -45,6 +50,14 @@ def test_wraps_py3():
 
     assert kwonly_non_roundtrippable_repr() == (
         True, 'kwonly_non_roundtrippable_repr', 2)
+
+
+def test_update_wrapper_partial():
+    wrapper = functools.partial(wrappable_varkw_func, b=1)
+    functools.update_wrapper(wrapper, wrapper.func)
+
+    fully_wrapped = update_wrapper(wrapper, wrappable_varkw_func, build_from=wrapper)
+    assert fully_wrapped(1) == (1, 1)
 
 
 def test_remove_kwonly_arg():
